improve AI consultation and divide models in patients in submodels folder
