+++ patients/submodels/__init__.py
@@ -1,0 +1,5 @@
+from .consultation_models import *
+from .dependent_models import *
+from .patient_models import *
+from .appointment_models import *
+from .utils import *
